+++ src/main/java/org/thehellnet/shab/protocol/line/HabPositionLine.java
@@ -12,9 +12,9 @@ public class HabPositionLine extends Line {
     private static final Command COMMAND = Command.HAB_POSITION;
     public static final String COMMAND_TAG = "HP";
 
-    private float latitude;
-    private float longitude;
-    private float altitude;
+    private double latitude;
+    private double longitude;
+    private double altitude;
 
     public HabPositionLine() {
         super(COMMAND);
@@ -40,27 +40,27 @@ public class HabPositionLine extends Line {
         altitude = Float.parseFloat(items[4]);
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public float getAltitude() {
+    public double getAltitude() {
         return altitude;
     }
 
-    public void setAltitude(float altitude) {
+    public void setAltitude(double altitude) {
         this.altitude = altitude;
     }
 }
